refactor(intents): split the 962-line service.go by concern
Pure same-package file reorganization of the intent-processing service (no
behavior change):

- service.go (962 -> 177): Service + ProcessIntent/processStart/processStop core
- intent_start_profile.go:    hardware + HWAccel + profile-spec resolution
- intent_requested_codecs.go: the requested-codec computation/clamp/merge cluster
- intent_start_session.go:    build/persist/publish the started session + admission

Build + go vet clean; intents tests green; golangci-lint --new-from-rev = 0; verify-config green.
